# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/users": {
            "get": {
                "security": [
                    {
                        "ServiceName": []
                    },
                    {
                        "ServiceKey": []
                    }
                ],
                "description": "Get a paginated list of users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "List users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by email",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by active status",
                        "name": "active",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.ListOutput"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ServiceName": []
                    },
                    {
                        "ServiceKey": []
                    }
                ],
                "description": "Create a new user with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.CreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.CreateOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "ServiceName": []
                    },
                    {
                        "ServiceKey": []
                    }
                ],
                "description": "Get user details by unique ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get a user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.GetOutput"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ServiceName": []
                    },
                    {
                        "ServiceKey": []
                    }
                ],
                "description": "Update user details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.UpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.UpdateOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ServiceName": []
                    },
                    {
                        "ServiceKey": []
                    }
                ],
                "description": "Soft delete a user by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.DeleteOutput"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_infrastructure_web_handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.CreateInput": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "email": {
                    "description": "Email (validado via binding + UseCase)",
                    "type": "string",
                    "maxLength": 255
                },
                "name": {
                    "description": "Nome do user",
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.CreateOutput": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Timestamp no formato RFC3339",
                    "type": "string"
                },
                "id": {
                    "description": "ID gerado (UUID v7)",
                    "type": "string"
                }
            }
        },
        "github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.DeleteOutput": {
            "type": "object",
            "properties": {
                "deleted_at": {
                    "description": "Timestamp da deleção",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.GetOutput": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.ListOutput": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.GetOutput"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.PaginationOutput"
                }
            }
        },
        "github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.PaginationOutput": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.UpdateInput": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "Email (opcional)",
                    "type": "string",
                    "maxLength": 255
                },
                "name": {
                    "description": "Nome (opcional)",
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "github_com_DenysonJ_financial-wallet_internal_usecases_user_dto.UpdateOutput": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "internal_infrastructure_web_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string",
                            "example": "error description"
                        }
                    }
                }
            }
        }
    }
}